+++ okio/src/nonJvmMain/kotlin/okio/Buffer.kt
@@ -216,11 +216,15 @@ actual class Buffer : BufferedSource, BufferedSink {
   actual override fun indexOf(b: Byte, fromIndex: Long): Long = indexOf(b, fromIndex, Long.MAX_VALUE)
 
   actual override fun indexOf(b: Byte, fromIndex: Long, toIndex: Long): Long =
-    commonIndexOf(b, fromIndex, toIndex)
+    commonIndexOf(b, fromIndex = fromIndex, toIndex = toIndex)
 
   actual override fun indexOf(bytes: ByteString): Long = indexOf(bytes, 0)
 
-  actual override fun indexOf(bytes: ByteString, fromIndex: Long): Long = commonIndexOf(bytes, fromIndex)
+  actual override fun indexOf(bytes: ByteString, fromIndex: Long): Long =
+    indexOf(bytes, fromIndex, Long.MAX_VALUE)
+
+  actual override fun indexOf(bytes: ByteString, fromIndex: Long, toIndex: Long): Long =
+    commonIndexOf(bytes, fromIndex = fromIndex, toIndex = toIndex)
 
   actual override fun indexOfElement(targetBytes: ByteString): Long = indexOfElement(targetBytes, 0L)